+++ NEWS.md
@@ -1,3 +1,12 @@
 # fpppXYspan 0.1.0
 
 * Added a `NEWS.md` file to track changes to the package.
+
+# fpppXYspan 0.1.1
+
+* Updated the DESCRIPTION to avoid starting it with "This package", "This 
+function", package name, title or similar.
+
+* Updated the DESCRIPTION to omit the lengthy explanation about the contributors.
+
+* Added a logical argument to the `xyspan` function to express or suppress error messages.